feat: dom & node runtime plugins
