+++ packages/runtime/src/core.ts
@@ -102,7 +102,15 @@ export class FederationHost {
       ],
       HTMLLinkElement | void
     >(),
-    loadEntry: new AsyncHook(),
+    loadEntry: new AsyncHook<
+      [
+        {
+          remoteInfo: RemoteInfo;
+          remoteEntryExports?: RemoteEntryExports;
+        },
+      ],
+      Promise<RemoteEntryExports>
+    >(),
     // only work for manifest , so not open to the public yet
     fetch: new AsyncHook<
       [string, RequestInit],
