+++ packages/sdk/src/dom.ts
@@ -22,20 +22,11 @@ export function isStaticResourcesEqual(url1: string, url2: string): boolean {
   return relativeUrl1 === relativeUrl2;
 }
 
-export type CreateScriptHookReturn =
-  | HTMLScriptElement
-  | { script?: HTMLScriptElement; timeout?: number }
-  | void;
-
 export function createScript(info: {
   url: string;
   cb?: (value: void | PromiseLike<void>) => void;
   attrs?: Record<string, any>;
   needDeleteScript?: boolean;
-  createScriptHook?: (
-    url: string,
-    attrs?: Record<string, any> | undefined,
-  ) => CreateScriptHookReturn;
 }): { script: HTMLScriptElement; needAttach: boolean } {
   // Retrieve the existing script element by its src attribute
   let script: HTMLScriptElement | null = null;
@@ -57,8 +48,12 @@ export function createScript(info: {
     script = document.createElement('script');
     script.type = 'text/javascript';
     script.src = info.url;
-    if (info.createScriptHook) {
-      const createScriptRes = info.createScriptHook(info.url, info.attrs);
+
+    //@ts-ignore
+    const loaderHooks = __webpack_require__.federation.instance.loaderHook;
+    const createScriptHook = loaderHooks.lifecycle.createScript;
+    if (createScriptHook) {
+      const createScriptRes = createScriptHook.emit(info.url, info.attrs);
 
       if (createScriptRes instanceof HTMLScriptElement) {
         script = createScriptRes;
@@ -205,13 +200,9 @@ export function loadScript(
   url: string,
   info: {
     attrs?: Record<string, any>;
-    createScriptHook?: (
-      url: string,
-      attrs?: Record<string, any> | undefined,
-    ) => CreateScriptHookReturn;
   },
 ) {
-  const { attrs = {}, createScriptHook } = info;
+  const { attrs = {} } = info;
   return new Promise<void>((resolve, _reject) => {
     const { script, needAttach } = createScript({
       url,
@@ -220,7 +211,6 @@ export function loadScript(
         fetchpriority: 'high',
         ...attrs,
       },
-      createScriptHook,
       needDeleteScript: true,
     });
     needAttach && document.head.appendChild(script);
